backend/requirements.txt
anthropic
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
httpx
